// src/app/models/docentes.model.ts
export interface Docentes{
    pebemplPidm?: number;
    idBanner?: string;
    numeroDocumento?: String;
    apellido?: string;
    nombres?: string;
    prefijo?: string;
    fechaNacimiento?: string;
    nacionalidad?: string;
    sexo?: string;
    tipoSangre?: string;
    estadoCivilSith?: string;
    campus?: string;
    departamento?: string;
    codeDedicacion?: string; 
    tipoEmpleado?: string;
    estado?:string;
    status?:string;
    dedicacion?:string;
}